requirements.txt
deepforest==1.2.7
fastapi==0.95.1
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.5.0
starlette==0.26.1
uvicorn==0.21.1
